requirements.txt
# Real-Time Social Media Trend Analyzer - Dependencies

# Core Pipeline
kafka-python==2.0.2
pyspark==3.5.1

# NLP & Sentiment
textblob==0.18.0
nltk==3.9.1
vaderSentiment==3.3.2

# Data Handling
pandas==2.2.2
pyarrow==16.1.0

# Dashboard & Visualization
streamlit==1.36.0
plotly==5.22.0
wordcloud==1.9.3
matplotlib==3.9.0

# Utilities
faker==25.8.0